src/test/javascript/spec/app/entities/rendez-vous/rendez-vous.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { RendezVousService } from 'app/entities/rendez-vous/rendez-vous.service';
import { IRendezVous, RendezVous } from 'app/shared/model/rendez-vous.model';

describe('Service Tests', () => {
  describe('RendezVous Service', () => {
    let injector: TestBed;
    let service: RendezVousService;
    let httpMock: HttpTestingController;
    let elemDefault: IRendezVous;
    let expectedResult: IRendezVous | IRendezVous[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(RendezVousService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new RendezVous(0, 'AAAAAAA', 'AAAAAAA', currentDate, currentDate, currentDate);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            debutRV: currentDate.format(DATE_TIME_FORMAT),
            finRV: currentDate.format(DATE_TIME_FORMAT),
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a RendezVous', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            debutRV: currentDate.format(DATE_TIME_FORMAT),
            finRV: currentDate.format(DATE_TIME_FORMAT),
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            debutRV: currentDate,
            finRV: currentDate,
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.create(new RendezVous()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a RendezVous', () => {
        const returnedFromService = Object.assign(
          {
            codePatient: 'BBBBBB',
            codePS: 'BBBBBB',
            debutRV: currentDate.format(DATE_TIME_FORMAT),
            finRV: currentDate.format(DATE_TIME_FORMAT),
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            debutRV: currentDate,
            finRV: currentDate,
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of RendezVous', () => {
        const returnedFromService = Object.assign(
          {
            codePatient: 'BBBBBB',
            codePS: 'BBBBBB',
            debutRV: currentDate.format(DATE_TIME_FORMAT),
            finRV: currentDate.format(DATE_TIME_FORMAT),
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            debutRV: currentDate,
            finRV: currentDate,
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a RendezVous', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
